Create BrutForceOperation and move brutForce function code into it.
